requirements/prod.txt
Django == 3.2.3
requests
djangorestframework
pandas_datareader